Introduced a CNI guard interface which is used by Windows to lock the cleanup method.
