+++ agent/ecscni/plugin_unsupported.go
@@ -20,7 +20,6 @@ import (
 	"errors"
 	"time"
 
-	"github.com/containernetworking/cni/libcni"
 	"github.com/containernetworking/cni/pkg/types/current"
 )
 
@@ -29,10 +28,12 @@ const (
 	vpcCNIPluginPath = "/log/vpc-branch-eni.log"
 )
 
-// cniClient is the client to call plugin and setup the network
-type cniClient struct {
-	pluginsPath string
-	libcni      libcni.CNI
+// newCNIGuard returns a new instance of CNI guard for the CNI client.
+// It is supported only on Windows.
+func newCNIGuard() cniGuard {
+	return &guard{
+		mutex: nil,
+	}
 }
 
 type cniPluginVersion struct{}
@@ -41,11 +42,6 @@ type cniPluginVersion struct{}
 // On unsupported platforms, we will return an error
 func (client *cniClient) setupNS(ctx context.Context, cfg *Config) (*current.Result, error) {
 	return nil, errors.New("unsupported platform")
-}
-
-// cleanupNS is called by CleanupNS to cleanup the task namespace by invoking DEL for given CNI configurations.
-func (client *cniClient) cleanupNS(ctx context.Context, cfg *Config) error {
-	return errors.New("unsupported platform")
 }
 
 // ReleaseIPResource marks the ip available in the ipam db
